+++ NEWS.md
@@ -1,4 +1,5 @@
 # Version 1.0.9000.9097
+* Renamed `stopList` slot in `StoppingList` class.
 * Added `unique` flag to `assert_probabilities` checkmate custom functions.
 * In `StoppingTargetBiomarker` class, changed `scale` character slot to
   `is_relative` flag.